+++ anyblok/config.py
@@ -102,19 +102,20 @@ def get_url(db_name=None):
     if url:
         url = make_url(url)
         if username:
-            url.username = username
+            url = url.set(username=username)
         if password:
-            url.password = password
+            url = url.set(password=password)
         if database:
-            url.database = database
+            url = url.set(database=database)
 
         return url
 
     if drivername is None:
         raise ConfigurationException('No driver name defined')
 
-    return URL(drivername, username=username, password=password, host=host,
-               port=port, database=database, query=query)
+    return URL.create(
+        drivername, username=username, password=password, host=host,
+        port=port, database=database, query=query)
 
 
 class ConfigurationException(LookupError):