+++ anyblok/tests/test_core_sqlbase.py
@@ -52,6 +52,12 @@ class TestCoreSQLBase:
         t1 = registry.Test.insert(id2=1)
         assert registry.Test.query().first() == t1
 
+    def test_insert_and_select(self, registry_declare_model):
+        Test = registry_declare_model.Test
+        t1 = Test.insert(id2=1)
+        assert Test.execute_sql_statement(
+            Test.select_sql_statement()).scalars().first() == t1
+
     def test_query_one_is_more_explicite(self, registry_declare_model):
         registry = registry_declare_model
         with pytest.raises(NoResultFound) as exc:
@@ -59,7 +65,7 @@ class TestCoreSQLBase:
 
         assert (
             str(exc._excinfo[1]) ==
-            "On Model 'Model.Test': No row was found for one()")
+            "On Model 'Model.Test': No row was found when one was required")
 
     def test_query_dictone_is_more_explicite(self, registry_declare_model):
         registry = registry_declare_model
@@ -68,7 +74,7 @@ class TestCoreSQLBase:
 
         assert (
             str(exc._excinfo[1]) ==
-            "On Model 'Model.Test': No row was found for dictone()")
+            "On Model 'Model.Test': No row was found when one was required")
 
     def test_multi_insert(self, registry_declare_model):
         registry = registry_declare_model
@@ -78,6 +84,31 @@ class TestCoreSQLBase:
         for x in range(nb_value):
             assert registry.Test.query().filter(
                 registry.Test.id2 == x).count() == 1
+
+    def test_classmethod_delete(self, registry_declare_model):
+        registry = registry_declare_model
+        nb_value = 3
+        Test = registry.Test
+        Test.multi_insert(*[{'id2': x} for x in range(nb_value)])
+        assert Test.query().count() == nb_value
+        id2 = 1
+        Test.execute_sql_statement(
+            Test.delete_sql_statement().where(Test.id2 == id2))
+        assert registry.Test.query().count() == nb_value - 1
+        assert registry.Test.query().filter(
+            registry.Test.id2 != id2).count() == nb_value - 1
+
+    def test_delete_by_query(self, registry_declare_model):
+        registry = registry_declare_model
+        nb_value = 3
+        registry.Test.multi_insert(*[{'id2': x} for x in range(nb_value)])
+        assert registry.Test.query().count() == nb_value
+        t = registry.Test.query().first()
+        id2 = t.id2
+        t.delete(byquery=True)
+        assert registry.Test.query().count() == nb_value - 1
+        assert registry.Test.query().filter(
+            registry.Test.id2 != id2).count() == nb_value - 1
 
     def test_delete(self, registry_declare_model):
         registry = registry_declare_model
@@ -126,12 +157,33 @@ class TestCoreSQLBase:
         t.flag_modified('id2')
         assert 'id2' in t.get_modified_fields()
 
+    def test_classmethod_update(self, registry_declare_model):
+        registry = registry_declare_model
+        nb_value = 3
+        Test = registry.Test
+        Test.multi_insert(*[{'id2': x} for x in range(nb_value)])
+        t = Test.query().filter_by(id2=1).one()
+        assert Test.execute_sql_statement(
+            Test.update_sql_statement().where(Test.id2 == 1).values(id2=100)
+        ).rowcount == 1
+        assert registry.Test.query().filter(
+            registry.Test.id2 == 100).first() == t
+
     def test_update(self, registry_declare_model):
         registry = registry_declare_model
         nb_value = 3
         registry.Test.multi_insert(*[{'id2': x} for x in range(nb_value)])
         t = registry.Test.query().first()
-        t.update(id2=100)
+        assert t.update(id2=100) == 1
+        assert registry.Test.query().filter(
+            registry.Test.id2 == 100).first() == t
+
+    def test_update_byquery(self, registry_declare_model):
+        registry = registry_declare_model
+        nb_value = 3
+        registry.Test.multi_insert(*[{'id2': x} for x in range(nb_value)])
+        t = registry.Test.query().first()
+        assert t.update(byquery=True, id2=100) == 1
         assert registry.Test.query().filter(
             registry.Test.id2 == 100).first() == t